Separar classes em arquivos diferentes
